# AI_Range/aigatewayservice/openapi/ai-gateway.yaml
openapi: 3.0.3
info:
  title: AI-Range AI Gateway Service
  version: 2.0.0
  description: >
    REST surface for the AI Gateway Service. Two independent surfaces live
    at the same /v1 base path, distinguished by which bearer token pool
    authenticates them (see securitySchemes below):

    - **Range automation** (`/recommendations`, `/assessments`, `/reports*`,
      `/admin/*`): the range control plane, the event-bus adapter (see
      src/eventAdapter.js), and the instructor dashboard call into the AI
      decision engine's four structured modes, authenticated with a named
      service token (`serviceAuth`).
    - **Student chat** (`/models`, `/chat/completions`, `/embeddings`): an
      OpenAI-compatible surface Open WebUI calls on behalf of a student,
      authenticated with a per-enclave tenant token (`chatAuth`) plus
      student identity forwarded in `X-OpenWebUI-User-*` headers. This is
      the only path allowed to reach the tower's vLLM instances, which bind
      loopback-only - see docs/range-inference-tower.html, "The shape of
      the system".
servers:
  - url: http://localhost:8080/v1
paths:
...
  /admin/approvals:
    get:
      summary: List the instructor approval queue
      description: >
        Recommendations eventAdapter.js marked pending_approval - a
        branch_change/extend_time, or anything below the confidence
        threshold. Defaults to status=pending_approval.
      operationId: listApprovals
      security: [{ serviceAuth: [] }, { chatAuth: [] }]
      parameters:
        - in: query
          name: exercise_id
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [pending_approval, approved, rejected] }
      responses:
        "200":
          description: Approval queue entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  approvals:
                    type: array
                    items: { $ref: "#/components/schemas/Approval" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/approvals/{id}:
    get:
      summary: Get one approval queue entry
      operationId: getApproval
      security: [{ serviceAuth: [] }, { chatAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Approval entry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Approval" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/approvals/{id}/approve:
    post:
      summary: Approve a pending recommendation
      operationId: approveApproval
      security: [{ serviceAuth: [] }, { chatAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Updated approval entry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Approval" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Already resolved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/approvals/{id}/reject:
    post:
      summary: Reject a pending recommendation
      operationId: rejectApproval
      security: [{ serviceAuth: [] }, { chatAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Updated approval entry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Approval" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Already resolved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/upstreams:
    get:
      summary: Report readiness for every configured model upstream
      operationId: listUpstreamStatus
      security: [{ serviceAuth: [] }, { chatAuth: [] }]
      responses:
        "200":
          description: Upstream status
          content:
            application/json:
              schema:
                type: object
                properties:
                  upstreams:
                    type: array
                    items: { $ref: "#/components/schemas/UpstreamStatus" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/quota/{studentId}:
    get:
      summary: Inspect a student's current quota usage
      operationId: getQuotaUsage
      security: [{ serviceAuth: [] }, { chatAuth: [] }]
      parameters:
        - in: path
          name: studentId
          required: true
          schema: { type: string }
        - in: query
          name: tenant
          schema: { type: string }
      responses:
        "200":
          description: Current usage and configured limits
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuotaUsage" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/audit:
    get:
      summary: Look up recent inference audit records
      operationId: listAuditEntries
      security: [{ serviceAuth: [] }, { chatAuth: [] }]
      parameters:
        - in: query
          name: tenant
          schema: { type: string }
        - in: query
          name: student_id
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: Audit entries, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items: { $ref: "#/components/schemas/AuditEntry" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/lora/load:
    post:
      summary: Hot-load a student LoRA adapter into the isolated LoRA upstream
      description: >
        See "The adaptation lab" in docs/range-inference-tower.html. Calls
        vLLM's /v1/load_lora_adapter on the configured LoRA base upstream,
        then registers the adapter in the model registry under `name` so it
        appears in GET /models.
      operationId: loadLoraAdapter
      security: [{ serviceAuth: [] }, { chatAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, path]
              properties:
                name: { type: string, description: "Served model name for the loaded adapter." }
                path: { type: string, description: "Filesystem path to the adapter, under /srv/airange/adapters." }
      responses:
        "200":
          description: Adapter loaded and registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  loaded: { type: string }
                  baseUrl: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
  /admin/lora/unload:
    post:
      summary: Unload a previously loaded student LoRA adapter
      operationId: unloadLoraAdapter
      security: [{ serviceAuth: [] }, { chatAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "200":
          description: Adapter unloaded and unregistered
          content:
            application/json:
              schema:
                type: object
                properties:
                  unloaded: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
components:
  securitySchemes:
    serviceAuth:
      type: http
      scheme: bearer
      description: >
        Named service token (range control plane, instructor dashboard) -
        config.services in src/config.js. AI_GATEWAY_API_KEY is the
        "control-plane" token, kept for backward compatibility with the
        single-shared-secret setup this started as.
    chatAuth:
      type: http
      scheme: bearer
      description: >
        Static per-enclave tenant token (config.tenants) identifying which
        Open WebUI instance a request came from. Student identity is
        additionally required via forwarded headers, set by Open WebUI when
        ENABLE_FORWARD_USER_INFO_HEADERS=true - not modeled here as OpenAPI
        parameters since a hand-written client never sets them directly:
          X-OpenWebUI-User-Id (required), X-OpenWebUI-User-Email,
          X-OpenWebUI-User-Name, X-OpenWebUI-User-Role.
  responses:
    BadRequest:
      description: Request failed schema validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing/malformed Authorization header
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Invalid API key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    BadGateway:
      description: The upstream model server returned an error or was unreachable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: invalid_request }
            message: { type: string }
          required: [code, message]
      required: [error]
...
    UpstreamStatus:
      type: object
      properties:
        id: { type: string, example: gpt-oss-120b }
        baseUrl: { type: string }
        servedModelName: { type: string }
        ready: { type: boolean }
...
    Approval:
      type: object
      description: >
        A recommendation eventAdapter.js routed to pending_approval instead
        of auto-applying - see src/store/approvalStore.js.
      properties:
        approval_id: { type: string, format: uuid }
        exercise_id: { type: string }
        student_id: { type: string }
        recommendation_type: { type: string }
        target_id: { type: string, nullable: true }
        confidence: { type: number }
        rationale: { type: string }
        reason: { type: string, description: "Why this went to the queue - see decideApproval() in eventAdapter.js." }
        prompt_version: { type: string }
        status: { type: string, enum: [pending_approval, approved, rejected] }
        created_at: { type: string, format: date-time }
        resolved_at: { type: string, format: date-time, nullable: true }
        resolved_by: { type: string, nullable: true }
    QuotaUsage:
      type: object
      properties:
        studentTokensThisHour: { type: integer }
        tenantTokensThisHour: { type: integer }
        activeStreams: { type: integer }
        limits:
          type: object
          properties:
            tokensPerHourPerStudent: { type: integer }
            tokensPerHourPerTenant: { type: integer }
            maxConcurrentStreamsPerStudent: { type: integer }
    AuditEntry:
      type: object
      description: One inference call - see src/audit.js.
      properties:
        tenant: { type: string }
        studentId: { type: string, nullable: true }
        mode: { type: string, enum: [chat, embedding, recommendation, assessment, capacity_analytics, report] }
        model: { type: string }
        promptTokens: { type: integer, nullable: true }
        completionTokens: { type: integer, nullable: true }
        latencyMs: { type: integer, nullable: true }
        finishReason: { type: string, nullable: true }
        promptVersion: { type: string, nullable: true }
        recorded_at: { type: string, format: date-time }